Put endpoint that uses from perfil.parlamentar in a separate url
